Revert "Schnorr sigs for Scala.js (#2784)"
This reverts commit a44c1f855063c0b9528a3292e82f9f05a43dccd3.
