+++ build.sbt
@@ -376,7 +376,10 @@ lazy val cryptoTest = crossProject(JVMPlatform, JSPlatform)
     name := "bitcoin-s-crypto-test",
     libraryDependencies ++= Deps.cryptoTest.value
   )
-  .dependsOn(crypto)
+  .dependsOn(
+    crypto,
+    testkitCore
+  )
 
 lazy val cryptoTestJVM = cryptoTest.jvm
