+++ crypto-test/.js/src/test/scala/org/bitcoins/crypto/RandomTest.scala
@@ -1,6 +1,9 @@
 package org.bitcoins.crypto
 
-class RandomTest extends BitcoinSCryptoTest {
+import org.scalatest.flatspec.AnyFlatSpec
+import org.scalatest.matchers.must.Matchers
+
+class RandomTest extends AnyFlatSpec with Matchers {
 
   it should "generate random bytes" in {
     val rnd = 1.to(16).map(_ => BCryptoCryptoRuntime.randomBytes(32))
