+++ crypto-test/.jvm/src/test/scala/org/bitcoins/crypto/AesCryptTest.scala
@@ -1,10 +1,12 @@
 package org.bitcoins.crypto
 
+import org.bitcoins.testkitcore.gen.{CryptoGenerators, NumberGenerator}
+import org.bitcoins.testkitcore.util.BitcoinSUnitTest
 import org.scalacheck.Gen
 import org.scalatest.compatible.Assertion
 import scodec.bits.{ByteVector, HexStringSyntax}
 
-class AesCryptTest extends BitcoinSCryptoTest {
+class AesCryptTest extends BitcoinSUnitTest {
   behavior of "AesEncrypt"
 
   val password = AesPassword.fromNonEmptyString("PASSWORD")
