+++ crypto-test/.jvm/src/test/scala/org/bitcoins/crypto/DERSignatureUtilTest.scala
@@ -1,12 +1,12 @@
 package org.bitcoins.crypto
 
+import org.bitcoins.core.util.NumberUtil
+import org.bitcoins.testkitcore.util.BitcoinSUnitTest
 import scodec.bits.ByteVector
-
-import scala.math.BigInt
 
 /** Created by chris on 3/23/16.
   */
-class DERSignatureUtilTest extends BitcoinSCryptoTest {
+class DERSignatureUtilTest extends BitcoinSUnitTest {
 
   val p2shSignature = ECDigitalSignature(
     "304402205b7d2c2f177ae76cfbbf14d589c113b0b35db753d305d5562dd0b61cbf366cfb02202e56f93c4f08a27f986cd424ffc48a462c3202c4902104d4d0ff98ed28f4bf8001")
@@ -37,30 +37,30 @@ class DERSignatureUtilTest extends BitcoinSCryptoTest {
   it must "retrieve the (r,s) values for a p2sh signature in bitcoin" in {
     val (r, s) = DERSignatureUtil.decodeSignature(p2shSignature)
     r must be(
-      CryptoNumberUtil.toBigInt(
+      NumberUtil.toBigInt(
         "5b7d2c2f177ae76cfbbf14d589c113b0b35db753d305d5562dd0b61cbf366cfb"))
     s must be(
-      CryptoNumberUtil.toBigInt(
+      NumberUtil.toBigInt(
         "2e56f93c4f08a27f986cd424ffc48a462c3202c4902104d4d0ff98ed28f4bf80"))
   }
 
   it must "retrieve the (r,s) values for a p2pkh signature in bitcoin" in {
     val (r, s) = DERSignatureUtil.decodeSignature(p2pkhSignature)
     r must be(
-      CryptoNumberUtil.toBigInt(
+      NumberUtil.toBigInt(
         "16ffdbb7c57634903c5e018fcfc48d59f4e37dc4bc3bbc9ba4e6ee39150bca03"))
     s must be(
-      CryptoNumberUtil.toBigInt(
+      NumberUtil.toBigInt(
         "119c2241a931819bc1a75d3596e4029d803d1cd6de123bf8a1a1a2c3665e1fac"))
   }
 
   it must "retrieve the (r,s) values from a p2pk signature in bitcoin" in {
     val (r, s) = DERSignatureUtil.decodeSignature(p2pkSignature)
     r must be(
-      CryptoNumberUtil.toBigInt(
+      NumberUtil.toBigInt(
         "0a5c6163f07b8d3b013c4d1d6dba25e780b39658d79ba37af7057a3b7f15ffa1"))
     s must be(
-      CryptoNumberUtil.toBigInt(
+      NumberUtil.toBigInt(
         "1fd9b4eaa9943f734928b99a83592c2e7bf342ea2680f6a2bb705167966b7420"))
   }
