+++ crypto-test/.jvm/src/test/scala/org/bitcoins/crypto/ECPublicKeyTest.scala
@@ -1,10 +1,12 @@
 package org.bitcoins.crypto
 
+import org.bitcoins.testkitcore.gen.CryptoGenerators
+import org.bitcoins.testkitcore.util.BitcoinSUnitTest
 import scodec.bits._
 
 import scala.concurrent.ExecutionContext
 
-class ECPublicKeyTest extends BitcoinSCryptoTest {
+class ECPublicKeyTest extends BitcoinSUnitTest {
 
   it must "be able to decompress keys" in {
     val uncompressed =
