+++ crypto-test/.jvm/src/test/scala/org/bitcoins/crypto/HashDigestTest.scala
@@ -1,8 +1,10 @@
 package org.bitcoins.crypto
 
+import org.bitcoins.testkitcore.gen.{CryptoGenerators, NumberGenerator}
+import org.bitcoins.testkitcore.util.BitcoinSUnitTest
 import scodec.bits._
 
-class HashDigestTest extends BitcoinSCryptoTest {
+class HashDigestTest extends BitcoinSUnitTest {
   behavior of "DoubleSha256Digest"
 
   it must "be constructable from 32 bytes" in {
