+++ crypto-test/.jvm/src/test/scala/org/bitcoins/crypto/JvmECPublicKeyTest.scala
@@ -1,7 +1,10 @@
 package org.bitcoins.crypto
 
+import org.bitcoins.testkitcore.gen.CryptoGenerators
+import org.bitcoins.testkitcore.util.BitcoinSUnitTest
+
 /** Public key tests specific to the JVM */
-class JvmECPublicKeyTest extends BitcoinSCryptoTest {
+class JvmECPublicKeyTest extends BitcoinSUnitTest {
 
   behavior of "JVMECPublicKeytest"
