+++ crypto-test/.jvm/src/test/scala/org/bitcoins/crypto/SignTest.scala
@@ -1,6 +1,9 @@
 package org.bitcoins.crypto
 
-class SignTest extends BitcoinSCryptoAsyncTest {
+import org.bitcoins.testkitcore.gen.CryptoGenerators
+import org.bitcoins.testkitcore.util.BitcoinSJvmTest
+
+class SignTest extends BitcoinSJvmTest {
 
   implicit override val generatorDrivenConfig: PropertyCheckConfiguration =
     generatorDrivenConfigNewCode
@@ -19,6 +22,17 @@ class SignTest extends BitcoinSCryptoAsyncTest {
       sigF.map { sig =>
         assert(pubKey.verify(hash.bytes, sig))
       }
+    }
+  }
+
+  it must "sign arbitrary pieces of data with arbitrary entropy correctly" in {
+    forAllAsync(CryptoGenerators.sha256Digest, CryptoGenerators.sha256Digest) {
+      case (hash, entropy) =>
+        val sigF = privKey.signWithEntropyFunction(hash.bytes, entropy.bytes)
+
+        sigF.map { sig =>
+          assert(pubKey.verify(hash.bytes, sig))
+        }
     }
   }
