+++ crypto/.js/src/main/scala/org/bitcoins/crypto/BCryptoCryptoRuntime.scala
@@ -4,9 +4,9 @@ import scodec.bits.ByteVector
 
 import java.math.BigInteger
 import scala.scalajs.js
-import scala.scalajs.js.JSStringOps._
 import scala.scalajs.js.typedarray._
-import scala.scalajs.js.{JavaScriptException, UnicodeNormalizationForm}
+import scala.scalajs.js.JSStringOps._
+import scala.scalajs.js.UnicodeNormalizationForm
 
 /** This is an implementation of [[CryptoRuntime]] that defaults to
   * Bcrypto (https://github.com/bcoin-org/bcrypto) when possible.
@@ -23,9 +23,7 @@ trait BCryptoCryptoRuntime extends CryptoRuntime {
   private lazy val sha1 = new SHA1
   private lazy val sha256 = SHA256Factory.create()
   private lazy val hmac = SHA512.hmac.apply().asInstanceOf[HMAC]
-
-  private lazy val ecdsa =
-    new ECDSA("SECP256K1", sha256, js.constructorOf[SHA256], null)
+  private lazy val ecdsa = new ECDSA("SECP256K1", sha256, sha256, null)
 
   private lazy val randomBytesFunc: Int => ByteVector =
     try {
@@ -151,56 +149,48 @@ trait BCryptoCryptoRuntime extends CryptoRuntime {
 
   override def tweakMultiply(
       publicKey: ECPublicKey,
-      tweak: FieldElement): ECPublicKey = {
-    val keyBuffer = ecdsa.publicKeyTweakMul(publicKey.bytes, tweak.bytes, true)
-    ECPublicKey.fromBytes(keyBuffer)
-  }
+      tweak: FieldElement): ECPublicKey = ???
 
-  def publicKeyConvert(buffer: ByteVector, compressed: Boolean): ByteVector =
-    publicKeyConvert(toNodeBuffer(buffer), compressed)
+  override def add(pk1: ECPrivateKey, pk2: ECPrivateKey): ECPrivateKey = ???
 
-  def publicKeyConvert(buffer: Buffer, compressed: Boolean): Buffer =
-    ecdsa.publicKeyConvert(buffer, compressed)
+  override def add(bytes: ByteVector, pk2: ECPrivateKey): ByteVector = ???
 
-  override def add(pk1: ECPublicKey, pk2: ECPublicKey): ECPublicKey = {
-    try {
-      val keyBuffer =
-        ecdsa.publicKeyCombine(js.Array(pk1.bytes, pk2.bytes), true)
-      ECPublicKey.fromBytes(keyBuffer)
-    } catch {
-      case ex: JavaScriptException =>
-        // check for infinity
-        val k1: Buffer =
-          if (pk1.isCompressed) pk1.bytes
-          else publicKeyConvert(pk1.bytes, compressed = true)
-
-        val k2: Buffer =
-          if (pk2.isCompressed) pk2.bytes
-          else publicKeyConvert(pk2.bytes, compressed = true)
-
-        if (
-          ((k1.head == 0x02 && k2.head == 0x03) ||
-            (k1.head == 0x03 && k2.head == 0x02)) &&
-          k1.tail == k2.tail
-        ) {
-          ECPublicKey.fromHex("00")
-        } else {
-          throw ex
-        }
-    }
-  }
+  override def add(pk1: ECPublicKey, pk2: ECPublicKey): ECPublicKey = ???
 
   override def pubKeyTweakAdd(
       pubkey: ECPublicKey,
-      privkey: ECPrivateKey): ECPublicKey = {
-    val keyBuffer = ecdsa.publicKeyTweakAdd(pubkey.bytes, privkey.bytes, true)
-    ECPublicKey.fromBytes(keyBuffer)
-  }
+      privkey: ECPrivateKey): ECPublicKey = ???
 
   override def isValidPubKey(bytes: ByteVector): Boolean =
     ecdsa.publicKeyVerify(bytes)
 
   override def isFullyValidWithBouncyCastle(bytes: ByteVector): Boolean = ???
+
+  override def schnorrSign(
+      dataToSign: ByteVector,
+      privateKey: ECPrivateKey,
+      auxRand: ByteVector): SchnorrDigitalSignature = {
+    val buffer = ecdsa.schnorrSign(dataToSign, privateKey.bytes) //, auxRand)
+    SchnorrDigitalSignature.fromBytes(buffer)
+  }
+
+  override def schnorrSignWithNonce(
+      dataToSign: ByteVector,
+      privateKey: ECPrivateKey,
+      nonceKey: ECPrivateKey): SchnorrDigitalSignature = ???
+
+  override def schnorrVerify(
+      data: ByteVector,
+      schnorrPubKey: SchnorrPublicKey,
+      signature: SchnorrDigitalSignature): Boolean = {
+    ecdsa.schnorrVerify(data, signature.bytes, schnorrPubKey.bytes)
+  }
+
+  override def schnorrComputeSigPoint(
+      data: ByteVector,
+      nonce: SchnorrNonce,
+      pubKey: SchnorrPublicKey,
+      compressed: Boolean): ECPublicKey = ???
 
   override def adaptorSign(
       key: ECPrivateKey,
@@ -261,22 +251,6 @@ trait BCryptoCryptoRuntime extends CryptoRuntime {
     require(accum.length == len,
             s"Need $len bytes for buffer -> bytevector conversion")
     ByteVector(accum.map(_.toByte))
-  }
-
-  override def decodePoint(bytes: ByteVector): ECPoint = {
-    if (bytes.size == 1 && bytes(0) == 0x00) {
-      ECPointInfinity
-    } else {
-      val decoded = ecdsa.curve
-        .applyDynamic("decodePoint")(toNodeBuffer(bytes))
-        .asInstanceOf[Point]
-
-      if (decoded.isInfinity())
-        ECPointInfinity
-      else
-        ECPoint(new BigInteger(decoded.getX().toString()),
-                new BigInteger(decoded.getY().toString()))
-    }
   }
 }
