+++ crypto/.js/src/main/scala/org/bitcoins/crypto/ECDSA.scala
@@ -11,7 +11,7 @@ import scala.scalajs.js.annotation._
 class ECDSA(
     name: String = "SECP256K1",
     hash: SHA256 = new SHA256,
-    xof: js.Dynamic = js.constructorOf[SHA256],
+    xof: SHA256 = new SHA256,
     pre: String = null)
     extends js.Object {
 
@@ -19,23 +19,9 @@ class ECDSA(
 
   def privateKeyVerify(key: Buffer): Boolean = js.native
 
-  def privateKeyTweakMul(key: Buffer, tweak: Buffer): Buffer =
-    js.native
-
   def publicKeyCreate(key: Buffer, compressed: Boolean): Buffer = js.native
 
   def publicKeyVerify(key: Buffer): Boolean = js.native
-
-  def publicKeyConvert(key: Buffer, compressed: Boolean): Buffer = js.native
-
-  def publicKeyTweakMul(key: Buffer, tweak: Buffer, compress: Boolean): Buffer =
-    js.native
-
-  def publicKeyTweakAdd(key: Buffer, tweak: Buffer, compress: Boolean): Buffer =
-    js.native
-
-  def publicKeyCombine(keys: js.Array[Buffer], compress: Boolean): Buffer =
-    js.native
 
   def sign(msg: Buffer, key: Buffer): Buffer = js.native
 
@@ -47,5 +33,10 @@ class ECDSA(
       param: Byte,
       compress: Boolean): Buffer = js.native
 
-  val curve: js.Dynamic = js.native
+  var schnorr: Schnorr = js.native
+
+  def schnorrSign(msg: Buffer, key: Buffer): Buffer = js.native
+
+  def schnorrVerify(msg: Buffer, sig: Buffer, key: Buffer): Boolean =
+    js.native
 }
