+++ crypto/src/main/scala/org/bitcoins/crypto/CryptoUtil.scala
@@ -195,9 +195,6 @@ trait CryptoUtil extends CryptoRuntime {
 
   override def sipHash(item: ByteVector, key: SipHashKey): Long =
     cryptoRuntime.sipHash(item, key)
-
-  override def decodePoint(bytes: ByteVector): ECPoint =
-    cryptoRuntime.decodePoint(bytes)
 }
 
 object CryptoUtil extends CryptoUtil
