+++ project/Deps.scala
@@ -1,4 +1,3 @@
-import Deps.{Compile, Test}
 import sbt._
 import org.portablescala.sbtplatformdeps.PlatformDepsPlugin.autoImport._
 
@@ -270,7 +269,8 @@ object Deps {
   def crypto: Def.Initialize[Seq[ModuleID]] = {
     Def.setting {
       List(
-        Compile.scodec.value
+        Compile.scodec.value,
+        Test.scalaTest.value
       )
     }
   }
@@ -294,8 +294,7 @@ object Deps {
   def cryptoTest = Def.setting {
     List(
       Test.scalaTest.value,
-      Test.scalacheck.value,
-      Compile.scalaTestPlus.value
+      Test.scalacheck.value
     )
   }
