+++ testkit-core/src/main/scala/org/bitcoins/testkitcore/util/BaseAsyncTest.scala
@@ -6,22 +6,22 @@ import org.bitcoins.core.protocol.blockchain.ChainParams
 import org.bitcoins.core.util.BitcoinSLogger
 import org.scalacheck.{Gen, Shrink}
 import org.scalactic.anyvals.PosInt
-import org.scalatest.concurrent.AsyncTimeLimitedTests
-import org.scalatest.flatspec.AsyncFlatSpec
-import org.scalatest.matchers.must.Matchers
-import org.scalatest.time.Span
 import org.scalatest.{
   Assertion,
   AsyncTestSuite,
   BeforeAndAfter,
   BeforeAndAfterAll
 }
+import org.scalatest.concurrent.AsyncTimeLimitedTests
+import org.scalatest.flatspec.AsyncFlatSpec
+import org.scalatest.matchers.must.Matchers
+import org.scalatest.time.Span
 import org.scalatestplus.scalacheck.ScalaCheckPropertyChecks
 
 import scala.annotation.nowarn
 import scala.collection.JavaConverters.iterableAsScalaIterableConverter
-import scala.concurrent.duration._
 import scala.concurrent.{ExecutionContext, Future}
+import scala.concurrent.duration._
 
 /** This is a base trait in bitcoin-s for async tests
   */
@@ -30,7 +30,8 @@ trait BaseAsyncTest
     with BeforeAndAfterAll
     with Matchers
     with ScalaCheckPropertyChecks
-    with AsyncTimeLimitedTests { this: AsyncTestSuite =>
+    with AsyncTimeLimitedTests
+    with BitcoinSLogger { this: AsyncTestSuite =>
 
   implicit def np: NetworkParameters = RegTest
 
@@ -290,10 +291,7 @@ trait BaseAsyncTest
   * uses the default scala execution context to run
   * the tests on
   */
-trait BitcoinSJvmTest
-    extends AsyncFlatSpec
-    with BaseAsyncTest
-    with BitcoinSLogger {
+trait BitcoinSJvmTest extends AsyncFlatSpec with BaseAsyncTest {
 
   implicit override def executionContext: ExecutionContext =
     scala.concurrent.ExecutionContext.global
